web_interface/requirements.txt
flask==2.3.3
flask-socketio==5.3.6
python-dotenv==1.0.0
openai==0.28.1
pandas==2.0.3
numpy==1.24.3
matplotlib==3.7.2
seaborn==0.12.2
requests==2.31.0
Pillow==10.0.1 